Fix error when trying to start a raw hook
Auto fix version does not work with raw hook.
hookArgs[0] of raw hook is "*" and 
hookArgs[1] of raw hook is "raw".
Looking for latest version of "*" results in undefined.
Therefore "raw != undefined is true and 
hookArgs[1] is set to undefined.
Resulting in an error that there is no version "undefined"